+++ src/airflow/main.py
@@ -1,6 +1,5 @@
 import argparse
 import os
-import sys
 from datetime import datetime, timedelta
 from pathlib import Path
